historical-date: Refactor URDateField to pull out `hdate` utilities for use by NodeTable and EdgeTables.
